Factor out AnnotationPoint/Span
